1040 delete given dataset version (#1063)
* fix bug with get/delete a specific version of dataset; frontend delete button is working

* dataset versioning styling

* public dataset versioning styling

* delete thumbnail and vis endpoint considering versioning

* might be a bug deleting thumbnails

* switch the order

* write utility funcitons to help manage delete

* backend delete is configured

* frontend delete wired in correctly

* delete modal color and prompt updated

* add docstrings

* add last modified

* add specific pytest

* add visualization tests to dataset versioning

* add more pytest after deleting the latest dataset

+++ backend/app/routers/thumbnails.py
@@ -8,6 +8,7 @@ from app.models.files import FileDB
 from app.models.thumbnails import (
     ThumbnailDB,
     ThumbnailDBViewList,
+    ThumbnailFreezeDB,
     ThumbnailIn,
     ThumbnailOut,
 )
@@ -68,8 +69,16 @@ async def remove_thumbnail(thumb_id: str, fs: Minio = Depends(dependencies.get_f
         ):
             dataset.thumbnail_id = None
             dataset.save()
-        fs.remove_object(settings.MINIO_BUCKET_NAME, thumb_id)
-        thumbnail.delete()
+
+        # if not being used in any version, remove the raw bytes
+        if (
+            await ThumbnailFreezeDB.find_one(
+                ThumbnailFreezeDB.origin_id == PydanticObjectId(thumb_id)
+            )
+        ) is None:
+            fs.remove_object(settings.MINIO_BUCKET_NAME, thumb_id)
+
+        await thumbnail.delete()
         return {"deleted": thumb_id}
     raise HTTPException(status_code=404, detail=f"Thumbnail {thumb_id} not found")
